requirements.txt
six==1.15.0
streamlit==0.89.0
opencv_python==4.5.3.56
numpy==1.19.5
tensorflow==2.6.0
Pillow==8.3.2
